% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAlignment.R
\name{calculate_gap_percentage}
\alias{calculate_gap_percentage}
\title{Calculate Gap Percentage in Alignment}
\usage{
calculate_gap_percentage(alignment)
}
\arguments{
\item{alignment}{AAMultipleAlignment object or MsaAAMultipleAlignment object}
}
\value{
Numeric value representing gap percentage
}
\description{
Helper function to calculate the percentage of gaps in the alignment
}
\references{
Alignment scoring, Gaps & Similarities. Sequence Alignment:
Scores, Gaps and Gap Penalties. [accessed 2024 Nov 5].
https://proteinstructures.com/sequence/sequence-alignment/
}
\keyword{internal}
