% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAlignment.R
\name{createAlignment}
\alias{createAlignment}
\title{Create Multiple Sequence Alignment}
\usage{
createAlignment(seq_data, method = c("MUSCLE", "ClustalW"))
}
\arguments{
\item{seq_data}{List containing sequences and metadata (output from readSequences function)}

\item{method}{Character, alignment method to use: "MUSCLE" (default) or "ClustalW"}
}
\value{
List containing:
  \itemize{
    \item alignment: The multiple sequence alignment object
    \item metadata: List with alignment statistics
  }
}
\description{
Performs multiple sequence alignment on protein sequences using either MUSCLE or
ClustalW algorithm with default settings. Takes processed sequences from
readSequences() function as input.
}
\examples{
# Read sequences: Dermcidin/Lacritin
Dermcidin_Lacritin_Sequences <- readSequences(system.file("extdata", "IPR043557_Dermcidin_Lacritin.fasta",
                                     package = "ProtConservR"))
# Create alignment using MUSCLE
alignment_muscle_DL <- createAlignment(Dermcidin_Lacritin_Sequences, method = "MUSCLE")
# Create alignment using ClustalW
alignment_clustal_DL <- createAlignment(Dermcidin_Lacritin_Sequences, method = "ClustalW")

}
\references{
Edgar, R.C. (2004) MUSCLE: multiple sequence alignment with high accuracy and
high throughput. \emph{Nucleic Acids Res.} 32(5):1792-1797.

Thompson, J.D., Higgins, D.G., Gibson, T.J. (1994) CLUSTAL W: improving the
sensitivity of progressive multiple sequence alignment through sequence weighting,
position-specific gap penalties and weight matrix choice.
\emph{Nucleic Acids Res.} 22(22):4673-4680.

Pagès H, Aboyoun P, Gentleman R, DebRoy S (2024). \emph{Biostrings: Efficient
manipulation of biological strings.} R package version 2.74.0,
\href{https://bioconductor.org/packages/Biostrings}{Link}.

Barry J. Grant, Ana P. C. Rodrigues, Karim M. ElSawy, J. Andrew McCammon, Leo S.
D. Caves, Bio3d: an R package for the comparative analysis of protein structures,
\emph{Bioinformatics}, Volume 22, Issue 21, November 2006, Pages 2695–2696,
\href{https://doi.org/10.1093/bioinformatics/btl461}{link}.

Bodenhofer U, Bonatesta E, Horejs-Kainrath C, Hochreiter S (2015).
“msa: an R package for multiple sequence alignment.” \emph{Bioinformatics},
31(24), 3997–3999. doi:10.1093/bioinformatics/btv494
}
