% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Dermcidin_Lacritin_AA_sequences}
\alias{Dermcidin_Lacritin_AA_sequences}
\title{Amino acid (AA) String Set containing Dermcidin and Lacritin AA Sequences From
Various Organisms.}
\format{
An AAStringSet object containing the following:
\describe{
  \item{sequences}{A collection of amino acid sequences for Dermcidin and Lacritin
  proteins from various organisms. Each sequence corresponds to a UniProt
  entry that matches the specified InterPro accession (IPR043557). (1) Dermcidin
  is an antibiotic peptide produced by sweat glands that is processed into a
  47-amino acid peptide, exhibiting antimicrobial activity against various
  pathogens. (2) Lacritin is a tear glycoprotein that stimulates tear secretion
  and supports epithelial health. It possesses antimicrobial, cytoprotective,
  and wound-healing properties.}
}
}
\source{
InterPro - EMBL-EBI. Available at: https://www.ebi.ac.uk/interpro/entry/InterPro/IPR043557/
}
\usage{
Dermcidin_Lacritin_AA_sequences
}
\description{
Amino acid (AA) String Set containing Dermcidin and Lacritin AA Sequences From
Various Organisms.
}
\examples{
\dontrun{
  data("Dermcidin_Lacritin_AA_sequences")
  # View the first few sequences
  head(Dermcidin_Lacritin_AA_sequences)
}
}
\keyword{datasets}
