% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Dermcidin_Lacritin_Alignment}
\alias{Dermcidin_Lacritin_Alignment}
\title{MSA Alignment Object For Dermcidin and Lacritin Protein Sequences}
\format{
List containing:
\describe{
 \item{alignment}{The multiple sequence alignment object.}
 \item{metadata}{List with alignment statistics.}
}
}
\source{
https://www.ebi.ac.uk/interpro/entry/InterPro/IPR043557/.
}
\usage{
Dermcidin_Lacritin_Alignment
}
\description{
Dermcidin and Lacritin Protein Sequences (anti-microbial proteins) taken from InterPro
}
\examples{
\dontrun{
 Dermcidin_Lacritin_Alignment
}
}
\keyword{datasets}
