% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Insulin_AA_sequences}
\alias{Insulin_AA_sequences}
\title{Amino acid (AA) String Set containing Insulin AA Sequences From Various Organisms}
\format{
An AAStringSet object containing the following:
\describe{
  \item{sequences}{A collection of amino acid sequences for insulin
  proteins from various organisms. Each sequence corresponds to a UniProt
  entry that matches the specified InterPro accession (IPR004825).}
}
}
\source{
InterPro - EMBL-EBI. Available at: https://www.ebi.ac.uk/interpro/entry/InterPro/IPR004825/
}
\usage{
Insulin_AA_sequences
}
\description{
Amino acid (AA) String Set containing Insulin AA Sequences From Various Organisms
}
\examples{
\dontrun{
  data("Insulin_AA_sequences")
  # View the first few sequences
  head(Insulin_AA_sequences)
}
}
\keyword{datasets}
