% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotInteractiveConservation.R
\name{plotInteractiveConservation}
\alias{plotInteractiveConservation}
\title{Visualize Protein Conservation Scores Interactively}
\usage{
plotInteractiveConservation(
  mapped_data,
  view_type = c("all", "heatmap", "line", "stats"),
  color_scheme = c("blues", "rainbow")
)
}
\arguments{
\item{mapped_data}{List containing mapped conservation scores (output from mapConservation())}

\item{view_type}{Character, type of visualization: "all" (default), "heatmap", "line", or "stats"}

\item{color_scheme}{Character, color scheme for conservation: "blues" (default) or "rainbow"}
}
\value{
plotly object containing the interactive visualization. On the
interactive map, the first plot is a heatmap, showing the conservation scores
across residues. The color legend on the right indicates that lighter colours
correspond to a larger conservation score, and darker colours correspond to a
lower conservation score. Hovering over the heatmap, we can see three values
x, y, and z. The relevant ones are x and z, where x represents the residue
position and z represents the associated conservation score. The interactive,
hoverable line plot below shows the same information in a different, line plot
type representation, indicating the residue name as well. At the very bottom
of the interactive map are some summary statistics of the max, mean, median,
and min conservation scores.
}
\description{
Creates an interactive visualization of protein conservation scores using plotly.
This function generates multiple visualization types including a heatmap view of
the conservation scores, a line plot showing conservation across residues, and
summary statistics.
}
\examples{
# Read and process data
sequences <- readSequences(system.file("extdata", "IPR043557_Dermcidin_Lacritin.fasta",
                          package = "ProtConservR"))
alignment <- createAlignment(sequences, method = "ClustalW")
conservation <- calculateConservation(alignment)
structure <- readStructure(system.file("extdata", "Dermcidin.pdb",
                          package = "ProtConservR"))
mapped_data <- mapConservation(conservation, structure)

# Create interactive visualization
plotInteractiveConservation(mapped_data)

}
\references{
Sievert C (2020). Interactive Web-Based Data Visualization with
R, plotly, and shiny. Chapman and Hall/CRC. ISBN 9781138331457, https://plotly-r.com.

R Core Team (2021). R: A language and environment for statistical
computing. R Foundation for Statistical Computing, Vienna, Austria.
URL https://www.R-project.org/.
}
