% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSequences.R
\name{readSequences}
\alias{readSequences}
\title{Read Protein Sequences from Fasta File}
\usage{
readSequences(input, format = "fasta")
}
\arguments{
\item{input}{Character, Path to multifasta file containing protein family sequences}

\item{format}{Character, input format (default: "fasta")}
}
\value{
List containing processed sequences and metadata
}
\description{
Reads in unaligned protein sequences provided in a FASTA file. Protein
sequences should be of a particular protein family, like hemoglobin, insulin,
etc. User can use sample data located in inst/extdata/, or they can pull FASTA
files of protein families of their choice from InterPro website as input into
this function.
}
\examples{
# Reading from file: Insulin fasta
Insulin_Sequences <- readSequences(system.file("extdata", "IPR004825_Insulin.fasta", package = "ProtConservR"))
head(Insulin_Sequences)

# Reading from file: Dermcidin/Lacritin fasta
Dermcidin_Lacritin_Sequences <- readSequences(system.file("extdata", "IPR043557_Dermcidin_Lacritin.fasta", package = "ProtConservR"))
head(Dermcidin_Lacritin_Sequences)

}
\references{
Pagès H, Aboyoun P, Gentleman R, DebRoy S (2024). \emph{Biostrings: Efficient
manipulation of biological strings.} R package version 2.74.0,
\href{https://bioconductor.org/packages/Biostrings}{Link}.

Barry J. Grant, Ana P. C. Rodrigues, Karim M. ElSawy, J. Andrew McCammon, Leo S.
D. Caves, Bio3d: an R package for the comparative analysis of protein structures,
\emph{Bioinformatics}, Volume 22, Issue 21, November 2006, Pages 2695–2696,
\href{https://doi.org/10.1093/bioinformatics/btl461}{link}.
}
