% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readStructure.R
\name{readStructure}
\alias{readStructure}
\title{Read Protein Structure from PDB File}
\usage{
readStructure(pdb_input)
}
\arguments{
\item{pdb_input}{Character. A file path to a PDB file.}
}
\value{
A list containing processed structural information, including atomic coordinates, chain IDs, and residue data.
}
\description{
This function reads and processes a protein structure from a specified PDB file.
Users must provide a file path to a PDB file.
}
\examples{
# Example with local PDB file: Insulin
Insulin_structure <- readStructure(system.file("extdata", "Insulin.pdb", package = "ProtConservR"))

# Example with local PDB file: Dermcidin/Lacritin
Dermcidin_Lacritin_structure <- readStructure(system.file("extdata", "Dermcidin.pdb", package = "ProtConservR"))

}
\references{
#' Barry J. Grant, Ana P. C. Rodrigues, Karim M. ElSawy, J. Andrew McCammon, Leo S.
D. Caves, Bio3d: an R package for the comparative analysis of protein structures,
\emph{Bioinformatics}, Volume 22, Issue 21, November 2006, Pages 2695–2696,
\href{https://doi.org/10.1093/bioinformatics/btl461}{link}.
}
