% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSequences.R
\name{validate_input_format}
\alias{validate_input_format}
\title{Validate Input Format}
\usage{
validate_input_format(input)
}
\arguments{
\item{input}{Character string to validate}
}
\value{
Logical indicating if input is valid
}
\description{
Validate input format for above function read_sequences
}
